Add new membership extension
