+++ .gitignore
@@ -460,3 +460,4 @@ healthchecksdb
 /DesktopModules/Vanjaro/URLManager/url-resources.zip
 /DesktopModules/Vanjaro/UXManager/Extensions/Apps/SectionSettings/apps-sectionsettings-resources.zip
 /DesktopModules/Vanjaro/UXManager/Extensions/Menu/GoogleReCaptcha/menu-integrations-googlerecaptcha-resources.zip
+/DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/menu-memberships-resources.zip
