+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/Components/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace Vanjaro.UXManager.Extensions.Menu.Memberships.Components
+{
+    public static class Constants
+    {
+        public const string LocalResourcesFile = "~/DesktopModules/Vanjaro/UXManager/Extensions/Menu/" + ExtensionInfo.Name + "/Views/App_LocalResources/Shared.resx";
+    }
+}
