+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/ExtensionInfo.cs
@@ -1,0 +1,10 @@
+﻿namespace Vanjaro.UXManager.Extensions.Menu.Memberships
+{
+    public class ExtensionInfo
+    {
+        public const string GUID = "335F614E-A66C-4154-B5EA-340C35CFC0A7";
+        public const string Name = "Memberships";
+        public const string FriendlyName = "Memberships";
+        public const string AssemblyName = "Vanjaro.UXManager.Extensions.Menu.Memberships";
+    }
+}
