+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/Resources.txt
@@ -1,0 +1,2 @@
+﻿Resources\
+Views\
