+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/Resources/Stylesheets/app.css
@@ -1,0 +1,8 @@
+﻿.vj-ux-manager.memberships-settings {
+    margin-top: 15px;
+}
+
+.col-md-3 {
+    flex: 0 0 25%;
+    max-width: 25%;
+}
