+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/Resources/UIEngine/AngularBootstrap/Layouts/Admin.htm
@@ -1,0 +1,3 @@
+﻿<div class="uilayout memberships">
+    [uilayout:Default]
+</div>
