+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Memberships/Views/App_LocalResources/Shared.resx
@@ -117,136 +117,13 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="DefaultAuthProvider.Text" xml:space="preserve">
-    <value>Default Authentication Provider</value>
+  <data name="Memberships.Text" xml:space="preserve">
+    <value>Memberships</value>
   </data>
-  <data name="Picture_DefaultFolder.Text" xml:space="preserve">
-    <value>Default Folder</value>
+  <data name="SearchKeywords.Text" xml:space="preserve">
+    <value>memberships,redirect after login,redirect after logout,require profile,valid profile,login,registeration,redirect after registeration,new user notification,user notification,registeration notification,email</value>
   </data>
-  <data name="General.Text" xml:space="preserve">
-    <value>General</value>
-  </data>
-  <data name="HideLoginControl.OFF" xml:space="preserve">
-    <value>No</value>
-  </data>
-  <data name="HideLoginControl.ON" xml:space="preserve">
-    <value>Yes</value>
-  </data>
-  <data name="HideLoginControl.Text" xml:space="preserve">
-    <value>Hide Login Control</value>
-  </data>
-  <data name="Login.Text" xml:space="preserve">
-    <value>Login</value>
-  </data>
-  <data name="LoginSettings.Text" xml:space="preserve">
-    <value>Login Setting</value>
-  </data>
-  <data name="Media.Text" xml:space="preserve">
-    <value>Media</value>
-  </data>
-  <data name="Pictures.Text" xml:space="preserve">
-    <value>Pictures</value>
-  </data>
-  <data name="Videos.Text" xml:space="preserve">
-    <value>Videos</value>
-  </data>
-  <data name="RedirectAfterLogin.Text" xml:space="preserve">
-    <value>Redirect After Login</value>
-  </data>
-  <data name="RedirectAfterLogout.Text" xml:space="preserve">
-    <value>Redirect After Logout</value>
-  </data>
-  <data name="RedirectAfterRegistration.Text" xml:space="preserve">
-    <value>Redirect After Registration</value>
-  </data>
-  <data name="Registeration.Text" xml:space="preserve">
-    <value>Registration</value>
-  </data>
-  <data name="SelectOption.Text" xml:space="preserve">
-    <value>None Specified</value>
-  </data>
-  <data name="SSL.Text" xml:space="preserve">
-    <value>SSL</value>
-  </data>
-  <data name="SSLEnabled.OFF" xml:space="preserve">
-    <value>No</value>
-  </data>
-  <data name="SSLEnabled.ON" xml:space="preserve">
-    <value>Yes</value>
-  </data>
-  <data name="SSLEnabled.Text" xml:space="preserve">
-    <value>SSL Enabled</value>
-  </data>
-  <data name="SSLEnforced.OFF" xml:space="preserve">
-    <value>No</value>
-  </data>
-  <data name="SSLEnforced.ON" xml:space="preserve">
-    <value>Yes</value>
-  </data>
-  <data name="SSLEnforced.Text" xml:space="preserve">
-    <value>SSL Enforced</value>
-  </data>
-  <data name="SSLOffloadHeaderValue.Text" xml:space="preserve">
-    <value>SSL Offload Header Value</value>
-  </data>
-  <data name="SSLSettings.Text" xml:space="preserve">
-    <value>SSL Settings</value>
-  </data>
-  <data name="SSLURL.Text" xml:space="preserve">
-    <value>SSL URL</value>
-  </data>
-  <data name="StandardURL.Text" xml:space="preserve">
-    <value>Standard URL</value>
-  </data>
-  <data name="SuccessfullyUpdated.Text" xml:space="preserve">
-    <value>Successfully updated</value>
-  </data>
-  <data name="Update.Text" xml:space="preserve">
-    <value>Update</value>
-  </data>
-  <data name="Picture_AllowableFileExtensions.Text" xml:space="preserve">
-    <value>Allowable File Extensions</value>
-  </data>
-  <data name="Picture_MaxUploadSize.Text" xml:space="preserve">
-    <value>Max Upload Size (MB)</value>
-  </data>
-  <data name="Video_AllowableFileExtensions.Text" xml:space="preserve">
-    <value>Allowable File Extensions</value>
-  </data>
-  <data name="Video_MaxUploadSize.Text" xml:space="preserve">
-    <value>Max Upload Size (MB)</value>
-  </data>
-  <data name="Video_DefaultFolder.Text" xml:space="preserve">
-    <value>Default Folder</value>
-  </data>
-  <data name="AsyncTimeout.Text" xml:space="preserve">
-    <value>Time Before Timeout (Seconds)</value>
-  </data>
-  <data name="AutoAccountUnlockDuration.Text" xml:space="preserve">
-    <value>Auto-Unlock Accounts After (Minutes)</value>
-  </data>
-  <data name="RememberCheckbox.Text" xml:space="preserve">
-    <value>Enable Remember Me on Login Control</value>
-  </data>
-  <data name="RememberCheckbox.OFF" xml:space="preserve">
-    <value>No</value>
-  </data>
-  <data name="RememberCheckbox.ON" xml:space="preserve">
-    <value>Yes</value>
-  </data>
-  <data name="FileExtensions.Text" xml:space="preserve">
-    <value>Allowable File Extensions</value>
-  </data>
-  <data name="MaxUploadSize.Text" xml:space="preserve">
-    <value>Max Upload Size (MB)</value>
-  </data>
-  <data name="HTMLPageHeader.Text" xml:space="preserve">
-    <value>HTML Page Header</value>
-  </data>
-  <data name="maxUploadSizeValidation.Text" xml:space="preserve">
-    <value>Maximum upload size must be between 12 and </value>
-  </data>
-  <data name="DefaultEndUserExtensionWhitelist.Text" xml:space="preserve">
-    <value>Default End User Extension Whitelist</value>
+  <data name="Settings.Text" xml:space="preserve">
+    <value>Settings</value>
   </data>
 </root>
