+++ DesktopModules/Vanjaro/UXManager/Extensions/Menu/Security/Controllers/SecurityController.cs
@@ -35,15 +35,6 @@ namespace Vanjaro.UXManager.Extensions.Menu.Security.Controllers
         {
             Dictionary<string, IUIData> Settings = new Dictionary<string, IUIData>();
 
-            var PageList = Library.Managers.PageManager.GetParentPages(PortalSettings.Current).Select(a => new { a.TabID, a.TabName, a.DisableLink });
-            dynamic Results = Managers.SecurityManager.GetRegistrationSettings(portalSettings);
-            dynamic basicLoginSettings = Managers.SecurityManager.GetBasicLoginSettings(portalSettings);
-            if (basicLoginSettings != null)
-            {
-                Settings.Add("UpdateBasicLoginSettingsRequest", new UIData { Name = "UpdateBasicLoginSettingsRequest", Options = basicLoginSettings.UpdateBasicLoginSettingsRequest });
-                Settings.Add("DefaultAuthProvider", new UIData { Name = "DefaultAuthProvider", Options = basicLoginSettings.AuthProviders, OptionsText = "Name", OptionsValue = "Value", Value = basicLoginSettings.Settings.DefaultAuthProvider });
-            }
-            Settings.Add("UserRegistration", new UIData { Name = "UserRegistration", Options = Results });
             Settings.Add("UpdateSslSettingsRequest", new UIData { Name = "UpdateSslSettingsRequest", Options = Managers.SecurityManager.GetSslSettings(portalSettings, userInfo) });
 
             List<TreeView> DefaultFolders = new List<TreeView>
@@ -76,20 +67,7 @@ namespace Vanjaro.UXManager.Extensions.Menu.Security.Controllers
             ActionResult actionResult = new ActionResult();
             try
             {
-                dynamic UpdateBasicLoginSettingsRequest = JsonConvert.DeserializeObject<dynamic>(settingData.UpdateBasicLoginSettingsRequest.ToString());
-                dynamic UpdateRegistrationSettingsRequest = JsonConvert.DeserializeObject<dynamic>(settingData.UpdateRegistrationSettingsRequest.ToString());
-                Entities.UpdateSslSettingsRequest UpdateSslSettingsRequest = JsonConvert.DeserializeObject<Entities.UpdateSslSettingsRequest>(settingData.UpdateSslSettingsRequest.ToString());
-                actionResult = UpdateBasicLoginSettings(UpdateBasicLoginSettingsRequest);
-                if (actionResult.HasErrors)
-                {
-                    return actionResult;
-                }
-
-                actionResult = UpdateRegistrationSettings(UpdateRegistrationSettingsRequest);
-                if (actionResult.HasErrors)
-                {
-                    return actionResult;
-                }
+                Entities.UpdateSslSettingsRequest UpdateSslSettingsRequest = JsonConvert.DeserializeObject<Entities.UpdateSslSettingsRequest>(settingData.UpdateSslSettingsRequest.ToString());              
 
                 actionResult = UpdateMediaSettings(settingData);
                 if (actionResult.HasErrors)
@@ -116,43 +94,7 @@ namespace Vanjaro.UXManager.Extensions.Menu.Security.Controllers
             }
             return actionResult;
         }
-        private ActionResult UpdateBasicLoginSettings(dynamic request)
-        {
-            ActionResult actionResult = new ActionResult();
-            if (!ModelState.IsValid)
-            {
-                foreach (System.Web.Http.ModelBinding.ModelState val in ModelState.Values)
-                {
-                    foreach (System.Web.Http.ModelBinding.ModelError err in val.Errors)
-                    {
-                        actionResult.AddError(HttpStatusCode.BadRequest.ToString(), err.ErrorMessage);
-                    }
-                }
-            }
-            try
-            {
-                if (actionResult.IsSuccess)
-                {
-                    int PortalId = PortalSettings.PortalId;
-                    string cultureCode = string.IsNullOrEmpty(request.CultureCode.ToString()) ? LocaleController.Instance.GetCurrentLocale(PortalId).Code : request.CultureCode.ToString();
-                    PortalInfo portalInfo = PortalController.Instance.GetPortal(PortalId);
-                    portalInfo.AdministratorId = Convert.ToInt32(request.PrimaryAdministratorId);
-                    PortalController.Instance.UpdatePortalInfo(portalInfo);
-                    PortalController.UpdatePortalSetting(PortalId, "DefaultAuthProvider", request.DefaultAuthProvider.ToString());
-                    PortalController.UpdatePortalSetting(PortalId, "Security_RequireValidProfileAtLogin", request.RequireValidProfileAtLogin.ToString(), false);
-                    PortalController.UpdatePortalSetting(PortalId, "Security_CaptchaLogin", request.CaptchaLogin.ToString(), false);
-                    PortalController.UpdatePortalSetting(PortalId, "Security_CaptchaRetrivePassword", request.CaptchaRetrivePassword.ToString(), false);
-                    PortalController.UpdatePortalSetting(PortalId, "Security_CaptchaChangePassword", request.CaptchaChangePassword.ToString(), false);
-                    PortalController.UpdatePortalSetting(PortalId, "HideLoginControl", request.HideLoginControl.ToString(), false);
-                }
-            }
-            catch (Exception exc)
-            {
-                actionResult.AddError(HttpStatusCode.InternalServerError.ToString(), exc.Message);
-            }
-            return actionResult;
-        }
-
+        
         private ActionResult UpdateGeneralSettings(dynamic request)
         {
             ActionResult actionResult = new ActionResult();
@@ -196,46 +138,7 @@ namespace Vanjaro.UXManager.Extensions.Menu.Security.Controllers
             }
             return actionResult;
         }
-
-        private ActionResult UpdateRegistrationSettings(dynamic request)
-        {
-            ActionResult ActionResult = new ActionResult();
-            PortalSettings portalSettings = PortalController.Instance.GetCurrentSettings() as PortalSettings;
-            try
-            {
-                bool.TryParse(request.UseEmailAsUsername.ToString(), out bool userEmailAsUsername);
-                if (userEmailAsUsername && UserController.GetDuplicateEmailCount() > 0)
-                {
-                    ActionResult.AddError("userEmailAsUsername", DotNetNuke.Services.Localization.Localization.GetString(Constants.ContainsDuplicateAddresses, Constants.LocalResourcesFile));
-                    return ActionResult;
-                }
-
-                string setting = request.RegistrationFields;
-                PortalInfo portalInfo = PortalController.Instance.GetPortal(portalSettings.PortalId);
-                portalInfo.UserRegistration = Convert.ToInt32(request.UserRegistration);
-                PortalController.Instance.UpdatePortalInfo(portalInfo);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_RegistrationFields", setting);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_RegistrationFormType", request.RegistrationFormType.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_UseEmailAsUserName", request.UseEmailAsUsername.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "EnableRegisterNotification", request.EnableRegisterNotification.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_UseAuthProviders", request.UseAuthenticationProviders.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_ExcludeTerms", request.ExcludedTerms.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_UseProfanityFilter", request.UseProfanityFilter.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_RequireUniqueDisplayName", request.RequireUniqueDisplayName.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Security_DisplayNameFormat", request.DisplayNameFormat.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Security_UserNameValidation", request.UserNameValidation.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Security_EmailValidation", request.EmailAddressValidation.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_RandomPassword", request.UseRandomPassword.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Registration_RequireConfirmPassword", request.RequirePasswordConfirmation.ToString(), true);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Security_RequireValidProfile", request.RequireValidProfile.ToString(), false);
-                PortalController.UpdatePortalSetting(portalSettings.PortalId, "Security_CaptchaRegister", request.UseCaptchaRegister.ToString(), false);
-            }
-            catch (Exception ex)
-            {
-                ActionResult.AddError("UpdateRegistrationSettings", ex.Message);
-            }
-            return ActionResult;
-        }
+        
         private ActionResult UpdateSslSettings(Entities.UpdateSslSettingsRequest request)
         {
             ActionResult ActionResult = new ActionResult();
