+++ Vanjaro Platform.sln
@@ -174,6 +174,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VanjaroMailProvider", "Desk
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleTagManager", "DesktopModules\Vanjaro\UXManager\Extensions\Menu\GoogleTagManager\GoogleTagManager.csproj", "{E2B4D291-CFEE-4A66-A169-C579FE006434}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Memberships", "DesktopModules\Vanjaro\UXManager\Extensions\Menu\Memberships\Memberships.csproj", "{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -595,6 +597,12 @@ Global
 		{E2B4D291-CFEE-4A66-A169-C579FE006434}.Internal|Any CPU.Build.0 = Release|Any CPU
 		{E2B4D291-CFEE-4A66-A169-C579FE006434}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E2B4D291-CFEE-4A66-A169-C579FE006434}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}.Internal|Any CPU.ActiveCfg = Internal|Any CPU
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}.Internal|Any CPU.Build.0 = Internal|Any CPU
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -678,6 +686,7 @@ Global
 		{0D642708-861A-4340-B7E9-1232AB0969BC} = {C06E9479-0ABA-4FA5-9489-2BE3E90167AB}
 		{27960AD4-D591-4A21-9D47-14323E0D10F5} = {0D642708-861A-4340-B7E9-1232AB0969BC}
 		{E2B4D291-CFEE-4A66-A169-C579FE006434} = {AE289E00-EFE5-4EB0-9AF1-9B7E406CBDA2}
+		{2E0ECA06-F6A2-4F74-A5CD-9BE4CA09B6E7} = {E801F63F-A7CB-4367-A310-BC2D89D4D027}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {54A96059-6996-42FC-B348-4E326F62E67A}
